backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
pymupdf
python-docx
google-cloud-aiplatform
google-genai
google-adk
pydantic
jinja2
